refactor: cleanup unit and integrations tests for account modules #58
