--- src/Identity.API/IAssemblyMarker.cs
@@ -1,3 +1,0 @@
-namespace Identity.API;
-
-public interface IAssemblyMarker { }
