+++ src/Identity/IAssemblyMarker.cs
@@ -1,0 +1,3 @@
+namespace Identity;
+
+public interface IAssemblyMarker { }
