+++ tests/Account.UnitTests/Account.UnitTests.csproj
@@ -25,8 +25,4 @@
       <ProjectReference Include="..\..\src\Account\Account.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="Application\Queries\GetPersonalInformation\" />
-    </ItemGroup>
-
 </Project>
