+++ tests/Identity.IntegrationTests/Identity.IntegrationTests.csproj
@@ -31,7 +31,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\Identity.API\Identity.API.csproj" />
+      <ProjectReference Include="..\..\src\Identity\Identity.csproj" />
     </ItemGroup>
 
 </Project>
